Config

Description

This module defines the options that can be passed to TorXakis.

Documentation

data Config

TorXakis configuration options.

Constructors

Config

Fields

smtLog :: Bool

Log all SMT commands?

availableSolvers :: Map SolverId SolverConfig

Available solvers that can be chosen from.

selectedSolver :: SolverId

getProc :: Config -> Maybe CreateProcess

Return a CreateProcess according to the selected solver in the configuration.

defaultConfig :: Config

TorXakis default configuration

newtype SolverId

Constructors

SolverId

Fields

solverId :: String

data SolverConfig

Constructors

SolverConfig

Fields

execName :: FilePath

Executable name.

smtArgs :: [String]

Arguments for to be passed to the smtSolver.

changeSolver :: Config -> String -> Config

Change the selected solver.

changeLog :: Config -> Bool -> Config

addSolvers :: Config -> Map SolverId SolverConfig -> Config

updateCfg :: Maybe a -> (Config -> a -> Config) -> Config -> Config